I just moved stuff around and added the comments
