+++ src/main/java/com/softwaredesign/project/staff/chefstrategies/ChefStrategy.java
@@ -1,7 +1,8 @@
 package com.softwaredesign.project.staff.chefstrategies;
 
 import java.util.List;
-import com.softwaredesign.project.order.Station;
+
+import com.softwaredesign.project.kitchen.Station;
 
 public interface ChefStrategy {
     Station chooseNextStation(List<Station> assignedStations);
